+++ samples/qt/src/Tests/tst_maintests.qml
@@ -1,8 +1,8 @@
 import QtQuick 2.0
 import QtTest 1.0
 
-import org.mozilla.Glean 1.2
-import generated 1.2
+import org.mozilla.Glean 1.3
+import generated 1.3
 
 import App 1.0